changed minor stuff in compliance with PEP8 and fixed request bugs and added better error handling
