--- openhivenpy/Types/__init__.py
@@ -1,43 +1,0 @@
-"""
-`openhivenpy.Types`
-
-Module Types 
-~~~~~~~~~~~~
-
-Module for all data classes that create Hiven Objects.
-
-MIT License
-~~~~~~~~~~~
-
-Copyright (c) 2020 Frostbyte Development Team 
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-"""
-
-from ._get_type import getType
-from .Room import *
-from .House import *
-from .Message import *
-from .User import *
-from .Context import *
-from .Member import *
-from .HivenClient import *
-from .Typing import *
-from .Attatchment import *
-from .Presence import *
